Divide wizard.cpp into several files. Move source files to src/. Add makefile. Delete compile.sh
