--- compile.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-clear
-echo "Compiling"
-echo "---------------"
-g++ -std=c++2a wizard.cpp -o wizard.out
