+++ src/functions.hpp
@@ -1,0 +1,5 @@
+#include <string>
+#include <regex>
+#include <iostream>
+void trimmer(std::string *arg);
+void abortApp();
