+++ src/help.cpp
@@ -1,0 +1,12 @@
+#include <iostream>
+void help(){
+    std::cout << "Usage: wizard [option]\n";
+    std::cout << "Available commands: \n";
+    std::cout << "\tconjure\tCreate a new package for a project\n";
+    std::cout << "\tspell\tCreate new component file\n";
+    std::cout << "\thelp\tDisplay list of commands and options\n";
+    std::cout << "\tlist\tLists supported programing languages\n";
+    //std::cout << "\tcharm\tInitializes directory as wizard project\n";
+    //std::cout << "\tsettings\n";
+    //std::cout << "\ttree\tProject tree\n";
+}
